+++ grr/client/grr_response_client/vfs_handlers/sleuthkit.py
@@ -40,7 +40,7 @@ class MyImgInfo(pytsk3.Img_Info):
     # Windows is unable to report the true size of the raw device and allows
     # arbitrary reading past the end - so we lie here to force tsk to read it
     # anyway
-    return long(1e12)
+    return 1e12
 
 
 class TSKFile(vfs.VFSHandler):
